# empresas_java/src/main/resources/api/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Está é uma documentação da api em Java",
    "version": "0.0.1",
    "title": "API Empresas",
    "contact": {
      "email": "qweasd@zxc.com"
    }
  },
  "host": "localhost:8080",
  "basePath": "/",
  "tags": [
    {
      "name": "empresas",
      "description": "Empresa para agenda de compromissos feita no desafio"
    }
  ],
  "schemes": [
    "http"
  ],
  "paths": {
    "/empresas": {
      "get": {
        "tags": [
          "empresas"
        ],
        "summary": "API que retorna os empresas cadastradas",
        "description": "Uma descrição de exemplo caso queira dar mais detahes sobre seu endpoint",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Sucesso",
            "schema": {
              "$ref": "#/definitions/ListaEmpresa"
            }
          },
          "401": {
            "description": "Token inválido",
            "schema": {
              "$ref": "#/definitions/RetornoStatusErro"
            }
          }
        },
        "security": [
          {
            "token_jwt_bearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "empresas"
        ],
        "summary": "Criando dados",
        "description": "Aqui iremos criar uma empresa",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Json enviado para cadastrar uma empresa",
            "required": true,
            "schema": {
              "$ref": "#/definitions/EmpresaEnvio"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Cadastrado com sucesso",
            "schema": {
              "$ref": "#/definitions/RetornoEmpresa"
            }
          },
          "400": {
            "description": "Os campos (id_usuario, titulo, data_inicio, data_fim) são obrigatórios",
            "schema": {
              "$ref": "#/definitions/RetornoStatusErro"
            }
          },
          "401": {
            "description": "Token inválido",
            "schema": {
              "$ref": "#/definitions/RetornoStatusErro"
            }
          }
        },
        "security": [
          {
            "token_jwt_bearer": []
          }
        ]
      }
    },
    "/Empresa/{id}": {
      "get": {
        "tags": [
          "empresas"
        ],
        "summary": "Retorna uma Empresa",
        "description": "Endpoint para retornar uma Empresa específico",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Id da Empresa",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "Sucesso",
            "schema": {
              "$ref": "#/definitions/RetornoEmpresa"
            }
          },
          "401": {
            "description": "Token inválido",
            "schema": {
              "$ref": "#/definitions/RetornoStatusErro"
            }
          },
          "404": {
            "description": "O id da Empresa enviado não existe",
            "schema": {
              "$ref": "#/definitions/RetornoStatusErro"
            }
          }
        },
        "security": [
          {
            "token_jwt_bearer": []
          }
        ]
      },
      "delete": {
        "tags": [
          "empresa"
        ],
        "summary": "Apagar um Empresa",
        "description": "Endpoint para apagar uma Empresa específica",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Id da Empresa",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "204": {
            "description": "Sem conteúdo"
          },
          "401": {
            "description": "Token inválido",
            "schema": {
              "$ref": "#/definitions/RetornoStatusErro"
            }
          },
          "404": {
            "description": "O id da Empresa enviado não existe",
            "schema": {
              "$ref": "#/definitions/RetornoStatusErro"
            }
          }
        },
        "security": [
          {
            "token_jwt_bearer": []
          }
        ]
      },
      "put": {
        "tags": [
          "empresas"
        ],
        "summary": "Atualizando dados",
        "description": "Aqui iremos atualizar uma Empresa específico",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Id da Empresa",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "in": "body",
            "name": "body",
            "description": "Json enviado para cadastrar uma Empresa",
            "required": true,
            "schema": {
              "$ref": "#/definitions/EmpresaEnvio"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Cadastrado com sucesso",
            "schema": {
              "$ref": "#/definitions/RetornoEmpresa"
            }
          },
          "401": {
            "description": "Token inválido",
            "schema": {
              "$ref": "#/definitions/RetornoStatusErro"
            }
          },
          "404": {
            "description": "O id da Empresa enviado não existe",
            "schema": {
              "$ref": "#/definitions/RetornoStatusErro"
            }
          },
          "400": {
            "description": "Os campos (id_usuario, titulo, data_inicio, data_fim) são obrigatórios",
            "schema": {
              "$ref": "#/definitions/RetornoStatusErro"
            }
          }
        },
        "security": [
          {
            "token_jwt_bearer": []
          }
        ]
      }
    }
  },
  "securityDefinitions": {
    "token_jwt_bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header" 
    }
  },
  "definitions": {
    "EmpresaEnvio": {
      "type": "object",
      "required": [
        "razao_social",
        "nome",
        "email",
        "cnpj",
        "dataCriacao",
        "dataAtualizacao"
      ],
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "razaoSocial":{
          "type": "string"
        },
        "nome": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "cnpj": {
          "type": "string"
        },
        "dataCriacao": {
          "type": "string"
        },
        "dataAtualizacao": {
          "type": "string"
        }
      }
    },
    "ListaEmpresa": {
      "type": "array",
      "items": {
        "$ref": "#/definitions/RetornoEmpresa"
      }
    },
    "RetornoEmpresa": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "razao_social":{
          "type": "string"
        },
        "nome": {
          "type": "string"
        },
        "cnpj": {
          "type": "string"
        },
        "data_criacao": {
          "type": "string"
        },
        "data_atualizacao": {
          "type": "string"
        }
      }
    },
    "RetornoStatusErro": {
      "type": "object",
      "properties": {
        "timestamp":{
          "type": "number"
        },
        "status":{
          "type": "number"
        },
        "error":{
          "type": "string"
        },
        "path":{
          "type": "string"
        }
      }
    }
  }
}